+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+- Add tornado to requirements file
+- Pass tx_id in call to process method
 
 ### 0.3.0
 - Move the data model to inheritence instead of composition